refactor: separate GasFee and SwapPath
